pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "seegull"
version = "0.1.0"
description = "Waste object detection models"
readme = "README.rst"
authors = [
  {name = "Bower", email = "hello@getbower.com"}
]
maintainers = [
  {name = "Bower", email = "hello@getbower.com"}
]
classifiers = [

]
license = {text = "MIT license"}
dependencies = [
  "autodistill>=0.1.26",
  "autodistill-grounding-dino>=0.1.3",
  "google-cloud-storage",
  "pillow_heif>=0.15.0",
  "tensorflow",
  "ultralytics",
  "ultralytics_bower @ git+https://git@github.com/PantaPasen/ultralytics.git",
]

[project.optional-dependencies]
dev = [
    "coverage",  # testing
    "mypy",  # linting
    "pytest",  # testing
    "ruff"  # linting
]

[project.urls]

bugs = "https://github.com/PantaPasen/seegull/issues"
changelog = "https://github.com/PantaPasen/seegull/blob/master/changelog.md"
homepage = "https://github.com/PantaPasen/seegull"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
"*" = ["*.*"]


[tool.coverage.run]
branch = true
# Fix for opencv
omit = ["config.py", "config-3.py"]

# Mypy
# ----

[tool.mypy]
files = "."

# Use strict defaults
strict = true
warn_unreachable = true
warn_no_return = true

[[tool.mypy.overrides]]
# Don't require test functions to include types
module = "tests.*"
allow_untyped_defs = true
disable_error_code = "attr-defined"

# Ruff
[tool.ruff]
line-length = 80
src = ["src"]